fix directory name
